backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
cryptography==44.0.0
redis==5.2.1
google-genai==1.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.7.1
python-multipart==0.0.20
numpy==2.2.2
email-validator==2.1.1
openai>=1.0.0
PyPDF2==3.0.1
python-docx==1.1.2
